move shielded params from core into shielded_token
