+++ crates/shielded_token/src/storage.rs
@@ -7,10 +7,11 @@ use namada_storage::{StorageRead, StorageWrite};
 use storage::ResultExt;
 
 use crate::storage_key::*;
+use crate::ShieldedParams;
 
 /// Initialize parameters for the token in storage during the genesis block.
 pub fn write_params<S>(
-    params: &token::MaspParams,
+    params: &ShieldedParams,
     storage: &mut S,
     address: &Address,
     denom: &token::Denomination,
@@ -18,7 +19,7 @@ pub fn write_params<S>(
 where
     S: StorageRead + StorageWrite,
 {
-    let token::MaspParams {
+    let ShieldedParams {
         max_reward_rate: max_rate,
         kd_gain_nom,
         kp_gain_nom,
