+++ crates/tests/src/integration/setup.rs
@@ -5,8 +5,8 @@ use std::str::FromStr;
 use std::sync::{Arc, Mutex};
 
 use color_eyre::eyre::{eyre, Result};
+use namada::token;
 use namada::types::dec::Dec;
-use namada::types::token;
 use namada_apps::cli::args;
 use namada_apps::client::utils::PRE_GENESIS_DIR;
 use namada_apps::config;
@@ -50,7 +50,7 @@ pub fn initialize_genesis() -> Result<(MockNode, MockServicesController)> {
     let mut templates = templates::All::read_toml_files(&template_dir)
         .expect("Missing genesis files");
     for (_, config) in templates.tokens.token.iter_mut() {
-        config.masp_params = Some(token::MaspParams {
+        config.masp_params = Some(token::ShieldedParams {
             max_reward_rate: Dec::from_str("0.1").unwrap(),
             kp_gain_nom: Dec::from_str("0.1").unwrap(),
             kd_gain_nom: Dec::from_str("0.1").unwrap(),
