+++ crates/token/src/lib.rs
@@ -15,7 +15,7 @@ use namada_storage::{Result, StorageRead, StorageWrite};
 
 /// Initialize parameters for the token in storage during the genesis block.
 pub fn write_params<S>(
-    params: &Option<MaspParams>,
+    params: &Option<ShieldedParams>,
     storage: &mut S,
     address: &Address,
     denom: &Denomination,
